fix: scripts move to js file
