+++ app/assets/less/main.less
@@ -9,3 +9,6 @@
 @text-white-color: #FFFFFF;
 @text-black-color: @theme-color;
 
+// value var
+@nav-height: 7vh;
+
